@@ -1,38 +1,41 @@
 source 'https://rubygems.org'
 
-gem 'rails', '~> 3.2'
-gem "nokogiri", '~> 1.5.6'
+gem 'rails', '4.2.2'
+
+gem "nokogiri", '~> 1'
 gem "sanitize", '~> 2.0.3'
 gem "paperclip"
-gem "rmagick"
-gem 'mysql2'
-
-gem "fastercsv", '~> 1.5.5'
+gem "pg"
 gem "haml", '~> 4'
 
 gem 'multi_json'
 
 gem "authlogic", :git => "git://github.com/binarylogic/authlogic.git"
 
-gem 'aws-sdk'
-gem 'cancan'
+gem 'aws-sdk', "~> 1.37"
+gem 'cancancan', '~> 1.10'
 gem 'annotate', ">=2.6.0"
 gem "sass", '~>3'
 gem 'stripe', :git => 'https://github.com/stripe/stripe-ruby'
 
+
+gem 'sass-rails'
+gem 'uglifier'
+gem 'coffee-rails'
+
 group :development do
-  gem "wirble"
-  gem "hirb"
-  gem 'sprockets'
-  gem 'uglifier'
+  gem 'capistrano', '>= 3.3.5'
+  gem 'capistrano-rvm'
+  gem 'capistrano-bundler' # for capistrano/bundler
+  gem 'capistrano-rails'
 end
 
-group :test do
-  gem 'factory_girl_rails'
-  gem "rspec-rails", ">= 2.12"
-  gem 'spork-rails'
-  gem 'capybara'
-  gem "capybara-webkit"
-  gem 'shoulda-matchers'
-  gem 'webmock', '~> 1.17'
-end
+# group :test do
+#   gem 'factory_girl_rails'
+#   gem "rspec-rails", ">= 2.12"
+#   gem 'spork-rails'
+#   gem 'capybara'
+#   gem "capybara-webkit"
+#   gem 'shoulda-matchers'
+#   gem 'webmock', '~> 1.17'
+# end
