@@ -9,11 +9,14 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20140323235222) do
+ActiveRecord::Schema.define(version: 20140323235222) do
 
-  create_table "tags", :force => true do |t|
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
+
+  create_table "tags", force: :cascade do |t|
     t.integer  "user_id"
     t.string   "name"
     t.string   "desc"
@@ -22,7 +25,7 @@ ActiveRecord::Schema.define(:version => 20140323235222) do
     t.datetime "updated_at"
   end
 
-  create_table "tconfigs", :force => true do |t|
+  create_table "tconfigs", force: :cascade do |t|
     t.integer  "user_id"
     t.string   "theme"
     t.string   "sort"
@@ -33,7 +36,7 @@ ActiveRecord::Schema.define(:version => 20140323235222) do
     t.datetime "updated_at"
   end
 
-  create_table "testimonials", :force => true do |t|
+  create_table "testimonials", force: :cascade do |t|
     t.integer  "user_id"
     t.integer  "tag_id"
     t.string   "token"
@@ -45,20 +48,20 @@ ActiveRecord::Schema.define(:version => 20140323235222) do
     t.text     "body"
     t.text     "body_edit"
     t.integer  "rating"
-    t.boolean  "publish",          :default => false
+    t.boolean  "publish",          default: false
     t.integer  "position"
-    t.boolean  "lock",             :default => false
+    t.boolean  "lock",             default: false
     t.string   "email"
     t.string   "meta"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "avatar_file_name", :default => "",    :null => false
-    t.boolean  "trash",            :default => false
+    t.string   "avatar_file_name", default: "",    null: false
+    t.boolean  "trash",            default: false
   end
 
-  create_table "theme_attributes", :force => true do |t|
+  create_table "theme_attributes", force: :cascade do |t|
     t.integer  "theme_id"
-    t.integer  "name",           :limit => 8
+    t.integer  "name",           limit: 8
     t.text     "published"
     t.text     "staged"
     t.text     "original"
@@ -67,18 +70,18 @@ ActiveRecord::Schema.define(:version => 20140323235222) do
     t.string   "attribute_name"
   end
 
-  create_table "themes", :force => true do |t|
+  create_table "themes", force: :cascade do |t|
     t.integer  "user_id"
-    t.integer  "name",       :limit => 8
-    t.boolean  "staged",                  :default => false
-    t.boolean  "published",               :default => false
+    t.integer  "name",       limit: 8
+    t.boolean  "staged",               default: false
+    t.boolean  "published",            default: false
     t.datetime "created_at"
     t.datetime "updated_at"
     t.string   "type"
     t.string   "theme_name"
   end
 
-  create_table "tweet_settings", :force => true do |t|
+  create_table "tweet_settings", force: :cascade do |t|
     t.integer  "user_id"
     t.string   "theme"
     t.string   "sort"
@@ -87,42 +90,41 @@ ActiveRecord::Schema.define(:version => 20140323235222) do
     t.datetime "updated_at"
   end
 
-  create_table "tweets", :force => true do |t|
+  create_table "tweets", force: :cascade do |t|
     t.integer  "user_id"
-    t.text     "data",       :limit => 16777215
+    t.text     "data"
     t.integer  "position"
-    t.boolean  "publish",                        :default => false
-    t.boolean  "trash",                          :default => false
+    t.boolean  "publish",    default: false
+    t.boolean  "trash",      default: false
     t.datetime "created_at"
     t.datetime "updated_at"
     t.string   "tweet_uid"
   end
 
-  create_table "users", :force => true do |t|
-    t.string   "email",                                  :null => false
-    t.string   "crypted_password",                       :null => false
-    t.string   "password_salt",                          :null => false
-    t.string   "persistence_token",                      :null => false
-    t.string   "single_access_token",                    :null => false
-    t.string   "apikey",                                 :null => false
-    t.integer  "login_count",         :default => 0,     :null => false
+  create_table "users", force: :cascade do |t|
+    t.string   "email",                               null: false
+    t.string   "crypted_password",                    null: false
+    t.string   "password_salt",                       null: false
+    t.string   "persistence_token",                   null: false
+    t.string   "single_access_token",                 null: false
+    t.string   "apikey",                              null: false
+    t.integer  "login_count",         default: 0,     null: false
     t.datetime "last_login_at"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "perishable_token",                       :null => false
-    t.boolean  "premium",             :default => false, :null => false
+    t.boolean  "premium",             default: false
     t.string   "stripe_id"
   end
 
-  add_index "users", ["email"], :name => "index_users_on_email"
-  add_index "users", ["persistence_token"], :name => "index_users_on_persistence_token"
+  add_index "users", ["email"], name: "index_users_on_email", using: :btree
+  add_index "users", ["persistence_token"], name: "index_users_on_persistence_token", using: :btree
 
-  create_table "widget_logs", :force => true do |t|
+  create_table "widget_logs", force: :cascade do |t|
     t.integer  "user_id"
     t.string   "url"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.integer  "impressions", :default => 0
+    t.integer  "impressions", default: 0
   end
 
 end
