@@ -5,9 +5,9 @@ class ApplicationController < ActionController::Base
     
   # See ActionController::RequestForgeryProtection for details
   # Uncomment the :secret if you're not using the cookie session store
-  protect_from_forgery # :secret => '1e2074d45c182ed5f3743d8d17ccd9c7'
-  
-  # See ActionController::Base for details 
+  #protect_from_forgery # :secret => '1e2074d45c182ed5f3743d8d17ccd9c7'
+
+  # See ActionController::Base for details
   # Uncomment this to filter the contents of submitted sensitive data parameters
   # from your application log (in this case, all fields with names like "password"). 
   # filter_parameter_logging :password
