@@ -25,10 +25,9 @@ module Seed
       :c_position   => "Founder",
       :url          => "worldunited.com",
       :location     => "Berkeley, CA",
-      :body         => 5,
       :body         => "The interface is simple and directed. I have a super busy schedule and did not want to waste any time learning yet another website. Pluspanda values my time. Thanks!",
       :publish      => 1
-    ) 
+    )
 
     self.testimonials.create(
       :name         => "John Doe",
@@ -36,18 +35,16 @@ module Seed
       :c_position   => "President",
       :url          => "supercompany.com",
       :location     => "Atlanta, Georgia",
-      :body         => 5,
       :body         => "The interface is simple and directed. I have a super busy schedule and did not want to waste any time learning yet another website. Pluspanda values my time. Thanks!",
       :publish      => 1
     )
-    
+
     self.testimonials.create(
       :name         => "Jane Smith",
       :company      => "Widgets R Us",
       :c_position   => "Sales Manager",
       :url          => "widgetsrus.com",
       :location     => "Los Angeles, CA",
-      :body         => 5,
       :body         => "Pluspanda makes our testimonials look great! Our widget sales our up 200% Thanks Pluspanda!",
       :publish      => 1
     )
@@ -57,25 +54,25 @@ end
 class User < ActiveRecord::Base
   include Seed
   acts_as_authentic
-  
+
   has_many :testimonials, :dependent => :destroy
-  has_many :standard_themes, :dependent => :destroy   
+  has_many :standard_themes, :dependent => :destroy
   has_one :tconfig, :dependent => :destroy
 
   has_many :widget_logs, :dependent => :destroy
-  
+
   validates_uniqueness_of :email
-  
+
   before_create :generate_defaults
   after_create  :create_dependencies
 
   attr_accessor :is_via_api
 
-  def generate_defaults 
+  def generate_defaults
     self.apikey = SecureRandom.hex(8)
   end
-  
-  
+
+
   def create_dependencies
     # standard testimonials
     create_tconfig
@@ -84,7 +81,7 @@ class User < ActiveRecord::Base
   end
 
   def deliver_password_reset_instructions!
-    reset_perishable_token!  
+    reset_perishable_token!
     UserMailer.password_reset_instructions(self.reload).deliver
   end
 
@@ -93,20 +90,20 @@ class User < ActiveRecord::Base
       SELECT count(id) as total_count, MONTH(created_at) as month, YEAR(`created_at`) as year
       FROM users
       GROUP BY month, year
-      HAVING month IS NOT NULL 
+      HAVING month IS NOT NULL
       ORDER BY year desc, month desc
     ")
   end
-  
+
   def self.total_by_login_count
     self.find_by_sql("
-      SELECT count(id) as total, if(login_count >= 8, 8, login_count) as logins 
+      SELECT count(id) as total, if(login_count >= 8, 8, login_count) as logins
       FROM users
       GROUP BY logins
       ORDER BY logins
     ")
   end
-  
+
   def self.stale
     self.find_by_sql("
       SELECT count(id) as total
@@ -115,7 +112,7 @@ class User < ActiveRecord::Base
       AND login_count <= 3
     ")
   end
-  
+
   def self.unstale
     self.find_by_sql("
       SELECT count(id) as total
@@ -124,5 +121,5 @@ class User < ActiveRecord::Base
       AND login_count > 3
     ")
   end
-  
+
 end
